# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Igloo API",
    "version": "0.1.0",
    "description": "HTTP API for the Igloo self-hosted media center. Most JSON endpoints return a shared envelope with error, optional message, and optional data fields."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin Igloo server"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/api/movies/{id}/subtitles/{trackIndex}/web.vtt": {
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "Extract a subtitle track as WebVTT",
        "operationId": "subtitleWebVTT",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          },
          {
            "$ref": "#/components/parameters/TrackIndexPath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/WebVTTResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/settings/": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get configured library paths",
        "operationId": "getSettings",
        "responses": {
          "200": {
            "$ref": "#/components/responses/SettingsResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/settings/general": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get general application settings",
        "description": "Admin-only endpoint.",
        "operationId": "getGeneralSettings",
        "responses": {
          "200": {
            "$ref": "#/components/responses/GeneralSettingsResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update general application settings",
        "description": "Admin-only endpoint.",
        "operationId": "updateGeneralSettings",
        "requestBody": {
          "$ref": "#/components/requestBodies/UpdateGeneralSettingsRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/UpdateGeneralSettingsResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/settings/scan/music": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Trigger a music library scan",
        "description": "Admin-only endpoint. The scan runs asynchronously.",
        "operationId": "triggerMusicScan",
        "responses": {
          "200": {
            "$ref": "#/components/responses/MessageSuccess"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/settings/scan/movies": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Trigger a movie library scan",
        "description": "Admin-only endpoint. The scan runs asynchronously.",
        "operationId": "triggerMovieScan",
        "responses": {
          "200": {
            "$ref": "#/components/responses/MessageSuccess"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Server-side session cookie set by POST /api/auth/login."
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 1
        }
      },
...
      "TrackIndexPath": {
        "name": "trackIndex",
        "in": "path",
        "required": true,
        "description": "Zero-based index into the movie subtitle rows ordered by stream_index.",
        "schema": {
          "type": "integer",
          "minimum": 0
        }
      },
...
    },
    "requestBodies": {
...
      "UpdateGeneralSettingsRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UpdateGeneralSettingsRequest"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "MessageSuccess": {
        "description": "Successful JSON response with an optional message.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageSuccess"
            }
          }
        }
      },
...
      "SettingsResponse": {
        "description": "Library path settings response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SettingsEnvelope"
            }
          }
        }
      },
      "GeneralSettingsResponse": {
        "description": "General settings response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GeneralSettingsEnvelope"
            }
          }
        }
      },
      "UpdateGeneralSettingsResponse": {
        "description": "Updated general settings response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UpdateGeneralSettingsEnvelope"
            }
          }
        }
      },
...
      "WebVTTResponse": {
        "description": "Extracted WebVTT subtitle payload.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-cache"
            }
          }
        },
        "content": {
          "text/vtt": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Bad request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid session.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated user is not allowed to perform this action.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "UnsupportedMediaType": {
        "description": "Media type cannot be converted or processed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "JsonSuccess": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "const": false
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": false
      },
      "MessageSuccess": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "const": false
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "const": true
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "SettingsData": {
        "type": "object",
        "required": [
          "music_dir",
          "movies_dir",
          "shows_dir"
        ],
        "properties": {
          "music_dir": {
            "type": [
              "string",
              "null"
            ]
          },
          "movies_dir": {
            "type": [
              "string",
              "null"
            ]
          },
          "shows_dir": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
      "SettingsEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SettingsData"
              }
            }
          }
        ]
      },
      "HardwareAccelerationDevice": {
        "type": "string",
        "enum": [
          "cpu",
          "apple",
          "nvidia",
          "intel"
        ]
      },
      "GeneralSettings": {
        "type": "object",
        "required": [
          "tmdb_key",
          "jellyfin_token",
          "spotify_client_id",
          "spotify_client_secret",
          "hardware_acceleration_device",
          "enable_logger",
          "enable_watcher",
          "download_images",
          "static_dir",
          "logs_dir",
          "server_upload_mbps"
        ],
        "properties": {
          "tmdb_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "jellyfin_token": {
            "type": [
              "string",
              "null"
            ]
          },
          "spotify_client_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "spotify_client_secret": {
            "type": [
              "string",
              "null"
            ]
          },
          "hardware_acceleration_device": {
            "$ref": "#/components/schemas/HardwareAccelerationDevice"
          },
          "enable_logger": {
            "type": "boolean"
          },
          "enable_watcher": {
            "type": "boolean"
          },
          "download_images": {
            "type": "boolean"
          },
          "static_dir": {
            "type": "string"
          },
          "logs_dir": {
            "type": "string"
          },
          "server_upload_mbps": {
            "type": [
              "number",
              "null"
            ]
          },
          "restart_required": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "UpdateGeneralSettingsRequest": {
        "type": "object",
        "required": [
          "tmdb_key",
          "jellyfin_token",
          "spotify_client_id",
          "spotify_client_secret",
          "hardware_acceleration_device",
          "enable_logger",
          "enable_watcher",
          "download_images",
          "static_dir",
          "logs_dir",
          "server_upload_mbps"
        ],
        "properties": {
          "tmdb_key": {
            "type": "string"
          },
          "jellyfin_token": {
            "type": "string"
          },
          "spotify_client_id": {
            "type": "string"
          },
          "spotify_client_secret": {
            "type": "string"
          },
          "hardware_acceleration_device": {
            "$ref": "#/components/schemas/HardwareAccelerationDevice"
          },
          "enable_logger": {
            "type": "boolean"
          },
          "enable_watcher": {
            "type": "boolean"
          },
          "download_images": {
            "type": "boolean"
          },
          "static_dir": {
            "type": "string"
          },
          "logs_dir": {
            "type": "string"
          },
          "server_upload_mbps": {
            "type": [
              "number",
              "null"
            ],
            "exclusiveMinimum": 0,
            "exclusiveMaximum": 100000
          }
        },
        "additionalProperties": false
      },
      "GeneralSettingsEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "settings"
                ],
                "properties": {
                  "settings": {
                    "$ref": "#/components/schemas/GeneralSettings"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
      "UpdateGeneralSettingsEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "settings",
                  "restart_required"
                ],
                "properties": {
                  "settings": {
                    "$ref": "#/components/schemas/GeneralSettings"
                  },
                  "restart_required": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
...
    }
  }
}